+++ src/main/java/co/ke/spsat/bowip/entities/ShoppingCart.java
@@ -30,7 +30,7 @@ public class ShoppingCart {
   //  @Column(name = "CARTITEM_ID", nullable = false)
     @JoinColumn(name = "CARTITEM_ID")//, referencedColumnName = "CARTITEM_ID")
     private List<CartItem> cartItems;
-    private Double totalAmount;
+    private BigDecimal totalAmount;
     private LocalDateTime creationDate;
     private boolean isActive;
